add literal test
